Throw exception for invalid template in dev mode
In developer mode php notices/warnings etc are promoted to being exceptions. This means the developer is a bit more diligent with cleaning up any messes as they occur.

I noticed this error on a Magento 1 client which is filled with "Not valid template file", I checked in M2 and this error seems to persist.

This change should make sure developers do not lead any dead templates in their layout.xml.