allows $this->getProduct to be null in product view block
From the following code we can find out there is a chance the result of
getProduct in \Magento\Catalog\Block\Product\View can be null:

~~~ php
/**
 * Retrieve current product model
 *
 * @return \Magento\Catalog\Model\Product | null
 */
public function getProduct()
{
    if (!$this->_coreRegistry->registry('product') && $this->getProductId()) {
        $product = $this->productRepository->getById($this->getProductId());
        $this->_coreRegistry->register('product', $product);
    }
    return $this->_coreRegistry->registry('product');
}
~~~

In the case there is no product in the registry and the getProductId
method returns something that gets interpreted as false, the product is
not fetched and also noet set in the registry. As a result we get the
product from the registry but its not there so the output of
getProduct() is null.

When we use the output of getProduct() further in the code without
checking we will eventually have breakage. This is now fixed by
returning an empty dummy product.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>