Make imageAsset protected to allow extensions to access the original file
I am a developer on a Magento 2 extension that optimizes images.

Since 2.1.6 there have been some changes to how product images are stored. I am trying to make the extension work with the latest changes, but the current implementation does not allow us to access the underlying `imageAsset` in order to retrieve/change the path of the images. This is required in order to make the extension work.

Please make `imageAsset` protected instead of private so we can update our extension and make it work with the latest Magento 2 versions again.