Update select.test.js
Added a scenario where options are loaded as an object instead of an array. Options are generated in PHP. This is the case when the option array in PHP has non-numeric keys or has numeric keys but misses some numeric keys.

I noticed the opts test was deleted by someone in an earlier commit, but I think this is too important to remove from the test, so I added it again.