Multi-threaded image resizing
We had the following scenario: A customer could only provide his
full catalog a day before the launch of his online store. All
image resizes had to be generated over the course of one night.
With a catalog of 100.000+ products and 3+ images per product
this was not possible with the current `catalog:images:resize`
command.

The customer did however have 4 24 core web servers at his
disposal which all shared a NFS `pub/media` drive. With this
addition to the `catalog:images:resize` command it is now
possible to run a 24 thread process which resizes 1/4 of the
product images on each web server. This way catalog images
are resized 96 times faster than before.

This also prevents the web shop from being way too slow and
way to heavy in it's requests in the first few days because
images are generated on the fly.

The code responsible for the php forks is heavily based on
the code used for static content deployment.