Change the order or run, schedule, cleanup to cleanup, schedule, run
This is in preparation of listening for changes in the configuration
to check if a scheduled cron job is still enabled. When you disable
a cron job in te config and flush the config cache, the cron observer
should remove any scheduled cron jobs that are pending. If the cron
observer runs the jobs before cleaning up, cron jobs that have been
disabled will still run 1 time before they are removed from the
schedule after disabling them.

The extra advantage of this setup is that cron jobs will run
directly after they have been scheduled, in the same run. So you
don't have to run the cron job twice to see jobs executing. One for
scheduling and one for running. If there is a perfectly good reason
not to generate the schedule and clean jobs before running them I
would very much like to know.