Fix Broken DHL Shipping Estimate
I may be doing something wrong, but it seems like after installing the DHL module and configuring the settings to connect our DHL account we were no longer able to get any shipping quotes on sale orders.  The root of the problem is that $request->getDestStreet returns an array for both Primary Address Fields.  In the existing code, the substr function fails when passing info to incov_substr (see error below) and the second function never returns a value.  I have changed it to index the array properly and only use the second value if it is set.

Error Given:
Warning: iconv_substr() expects parameter 1 to be string, array given in /var/www/magento2/vendor/magento/framework/Stdlib/StringUtils.php on line 183