Add a payload extender to the default shipping-save-processor
This will allow third party extensions to modify the payload for the
shipping address selection process, with the goal being the easy
addition of extension_attributes with as few extension conflicts as
possible.

By separating this out into it's own model (as opposed to including it
in the result of the processor return), non-default processors will also
be able to utilize it.