Check cart rule subselect conditions against quote item children too
The subselect condition only checked the visible quote items and this
proved to be a problem in the case of configurable and bundle products.
The quote item children are now checked against the validation too,
and an item will be considered valid and added to the subselect total
if either it, or at least one of it's children is validated.

In the case of bundle products, the children items data will be used
and added to the subselect total, when the match is on a child item.

In the case of configurable products, the parent item data will be used
in the subselect total, just like for all the other product types.

Resolves: #10477