FIX Strict-Transport-Security bug in magento backend
this commit fixes the infinite redirect loop in magento administration when user tries to update from non http to https.

default magento behavior is that if user opens https connection, it uses the https stream. If magento administration does not contain any configuration wheater to use https, only http is used, and redirection by browser does inifinite redirection. If magento config does not setup strict https (because with upgrade the https is not enabled at first) magento redirects administrator from https stream to http which is Strict-Transport-Security bug.

this PR fixes issue #11350 and  #11356 

please do not let @orlangur to process this one, because he believes that inifinite redircts in magento are not an issue, and he believes that magento behavior against Strict-Transport-Security is not an issue. Please note that it is severe issue, and should be fixed ASAP.