UPS Option to include TAX in rate
The current UPS Module uses the UPS Rate excluding TAX to use as the rate charged to the customer during checkout.

This modification adds an additional config option so you can choose whether to request and include TAX information from UPS in the rate or not.

This modification is especially useful for businesses who are not TAX registered and cannot claim back the TAX charged by UPS. This allows them to pass on the TAX costs to the customer as part of the overall shipping rate.