Suppressed Excessive Method Length checks
This pull request adds enough lines of code to the function to trigger the excessive method length check in PHPMD.

The majority of the code in the function is not mine and I do not want to change stuff for the sake of passing this check.

My goal was to make the changes with as little impact on the existing code base as possible.