Translate cache doesn't always fetch correct data
Their is an issue when Magento emulates another store localisation used is not always correct. This happens because the setConfig is populated with new data. But the cache ID is never cleared. 

This was discovered while testing Magento's checkout when Magento sends out order emails with a 3rd party extension.