Add configurable Monolog configuration
The Magento Log setup is made so that it is similar to Magento 1. In the
cases where you want to extend this functionality or change the way
logging is configured you have to write a custom module.

This change makes it possible to add custom log channels and
configuration in env.php without adding an extra module.