Fix for missing context reset of ComponentRegistrar
This test tries to setup a test environment with two modules and one
theme to test the generator.
During the test run, the test environment and the complete module stack
is processed by the generator.
So the test takes on my machine 25 seconds with xdebug enabled.
After resetting the registrar the complete test runs in ~50ms with
xdebug enabled.
Without the reset ~12.000 files are scanned. With the reset only 5 files
are scanned.

This test is not a unit test. It is an integration test.