Changed logic to obtain tax information from response to get rid of maths.
Previously the routine obtained the base cost, and then if taxes existed, it added the tax on to get the total.

Now the function checks for the presence of the "TotalChargesWithTaxes" field. If present and the "Request Tax-Inclusive Rate" option is set, then the total is obtained from the "TotalChargesWithTaxes" field. If not present, or the "Request Tax-Inclusive Rate" option is not set, then the total is obtained from the original field which does not include tax.

Also moved the setting of the "responseCurrencyCode" variable into the same logic. This ensures that the correct response currency code is always set correctly based on the value returned from UPS. Previously "responseCurrencyCode" would only have ever been set correctly if negotiated rates were never used.