Fix duplicate URL key error output
Message was;
`Url key: '%s' was already generated for an item with the SKU: '%s'. You need to specify the unique URL key manually in row(s): 1124`

Fixed message;
`3: Url key: 'venotrain-impuls' was already generated for an item with the SKU: 'Venotrain Impuls'. You need to specify the unique URL key manually in row(s): 1124`