magento/magento2#7816: Customer_account.xml file abused
This is in addiiton to magento/magento2#12852 commit 36cac17 - that fix
removed the title from being set in the `customer_account.xml` file but
the fix seemed to do nothing as the title was always being set in the
controller anyway. I've removed from the controller and added to the
correct layout file.