Optimize ID to SKU lookup of tier prices.
Previously, with a large number of tier or group prices, each tier would separately
make a database query to lookup the associated SKU.

This instead load the ID to SKU mapping once, and uses it for all tiers.

When lookup up a batch of 100 SKUs with hundreds of tier prices per SKU, this cuts
lookup time from on the order of 4s to on the order of 0.1s or less.