[2.2-develop] Update Gallery Template to handle boolean config Variables
Due to changes implemented in resolution to #12285, boolean configuration variables are now booleans, instead of the strings "true" and "false".

Previous to this fix, values that were true were being output as var : 1 and values that were false were being output as var :

This output for the false value is invalid.