Fix typehint of customer-data updateSectionId parameters
The typehint wrongly indicates the parameter should be a number, where
in fact it is a boolean.
It is passed through to the server side to the controller action
\Magento\Customer\Controller\Section\Load::execute()
where it is cast to a boolean and passed as the $updateIds parameter to
\Magento\Customer\CustomerData\SectionPool::getSectionsData()
and from there as the boolean parameter $forceUpdate to the
\Magento\Customer\CustomerData\Section\Identifier::initMark() method.

This PR introduces no backward incompatibilities, it only fixes the type
hint to make the code more readable and improve compatibility with static
code analysis tools.