Update the processing of items for the top menu
The counter in the foreach loop was not being calculated correctly. This
resolved in the 'last' css class not being set on the output when one or
more parent items are not active.

With this commit the items will be filtered first before they are being
processed for the html output.

Move the filtering to separate method

Moved the filter of children without active parent to a seperate method.
This will make the whole method more readable and understandable.

Also this is a little optimization; the foreach loop will only be run if
the child level = 0.

Apply consistent class import for Node.

Created private functions for the _getChildHtml method to reduce the
complexity for this function.