Fix catalog product attribute update webapi
If an existing attribute should be saved via webapi,
an exception "%1 already exists" is thrown.

The webapi endpoint "/rest/all/V1/products/attributes/{attributeCode}"
provides already an attribute code.

The save method checks if a "attributeId" is provided.
The use case of the "attributeCode" URL path paramter is to omit the
"attributeId" which is usally not known on API client side.

The fix replaces the "getAttributeId" method with "getAttributeCode".
That always works, because the "getAttributeCode" is used to load the
object by code in the next lines in the repository save method.