increase length of sequence_table column in sales_sequence_meta table
Having a too small length for the sequence_table column can lead to integrity constraint errors on setup:upgrade, because longer table names will be cut off at least the store id suffix of those sales sequence table names.
In our case it was the postfinance extension, which lead to the table names "sequence_postfinancecw_transaction_0" and "sequence_postfinancecw_transaction_1", which are 36 characters long.
So the name got cut off and as a consequence lead to an integrity constrained, when the recurring script in the Magento_SalesSequence module tries to add the meta data for the sequence_postfinancecw_transaction_1 table.

So a character length of only 32 is definitely not long enough. I guess 50 character should suffice though. ¯\_(ツ)_/¯