Make it possible to disable report bugs link
As developer I have removed this link a lot of times with a layout
update (remove=true).

Since we most of the times do not want this by default it should be
configurable whether to show this link or not.

A new setting is added under Design > Configuration > Edit > Other
Settings > Footer. The new label is 'Display Report Bugs Link'.

The default value for this setting is 'Yes' to keep it backwards
compatible.