Attribute with ID: "manufacturer" does not exist
When running the command `bin/magento setup:upgrade` to update to Magento 2.2.6, it failed for me with an error `Attribute with ID: "manufacturer" does not exist`. This is entirely correct: I have deliberately removed the attribute `manufacturer` from my shop. But the setup script should not fail for this.

This PR adds a simple `if`-check to see if the `manufacturer` attribute is there, before running the rest of the setup.

Note: In Magento 2.3, the entire procedure seems to have changed and this issue does not appear to be there anymore.