Fixes #18348 - Defaults not persisted for swatches.
When saving an attribute of type swatchtext or swatchvisual the selected
default was ignored and the last option always selected. This was due to
the JS always submitting all radio fields.

The standard dropdown attribute which this is built on top of only
submitted radio / checkbox fields if they were checked, so I copied this
behaviour forward.