Add MAGE_RUN_TYPE+MAGE_RUN_CODE to FPC identifier
Currently Magento 2.2.6 full page cache does not work correctly, when using SetEnv or similar techniques on unchanged URLs
So the MAGE_RUN_TYPE and MAGE_RUN_CODE should be added to the FPC identifier to allow for such use cases.
A common case is Magento 2 behind a reverse NGINX proxy.

This commit adds the MAGE_RUN_TYPE and MAGE_RUN_CODE to the cache idenfier, if present.
It is done inline with the design exceptions. Also the existing cache idenfier is change a little to accomondate
the new information properly.

Of course this will invalidate existing page caches with design exceptions, but as for any deploy the page
cache usually is invalidated, this is not a problem.