Updated label for Content Type configuration field.
Changed Allowable Methods (Non Doc and Doc) to always be displayed no
matter what the content type mode is.

Changed the isDutiable function to only look at whether a shipment is
domestic or not. Previously it would return dutiable for everything
if the Content Type mode is non document (which causes problems for
domestic shipments in Non Document mode). Fixes #19485