Prevent display of token when save for later is not selected
This sets the token `isVisible` value to faluse if
`VaultConfigProvider::IS_ACTIVE_CODE` is not set in the payment's
`additionalInformation` property. The value is not set when placing orders
through the admin panel, unless the save for later checkbox is selected. This
caused all admin order payment to be visible in the stored payment method
section in the storefront.

Fixes magento/magento2#19515

Testing on 2.3 depends on magento/magento2#19764