Fix region_id update in updateCustomerAddress mutation
This commit maps `region/region_id` to `AddressInterface:regionId` when
provided as input.

The `AddressInterface` contains both `region` and `regionId` properties which
are mapped to `\Magento\Customer\Model\Address` by the
`AddressRepositoryInterface`.

After populating an existing customer address with the user input for `region`,
the exting `regionId` was still set on the `AddressInterface`. When saving
the `region` property is mapped to the model before `regionId`.

https://github.com/magento/magento2/blob/1a92f1c1609445c7b6b107a165d387e306f2f779/app/code/Magento/Customer/Model/Address.php#L141-L167